重构storage

+++ dsdlsdk/storage.py
@@ -1,10 +1,12 @@
+import os.path
 import re
 import shutil
 from abc import ABC, abstractmethod
 from typing import Any, Dict, Tuple
 
 from boto3 import Session
-from commons.exceptions import CLIException, ExistCode
+
+from dsdlsdk.exception.exception import DatasetPathNotExists
 
 
 class Storage(ABC):
@@ -25,6 +27,8 @@ class Storage(ABC):
 class LocalDiskStorage(Storage):
 
     def remove_tree(self, local_path: str) -> bool:
+        if not os.path.exists(local_path):
+            raise DatasetPathNotExists(f"local path {local_path} not exist")
         try:
             shutil.rmtree(local_path, ignore_errors=True)
             return True
@@ -104,60 +108,3 @@ class SftpStorage(Storage):
 
     def __init__(self):
         raise NotImplementedError
-
-
-class StorageBuilder(object):
-
-    @staticmethod
-    def build_by_name(storage_name: str, config: Dict) -> Storage:
-        """
-        Build storage client by name
-        Args:
-            storage_name:
-            config:
-
-        Returns:
-
-        """
-        if "storage" in config.keys():
-            storage_config = config['storage'].get(storage_name, None)
-            if storage_config is not None:
-                path = storage_config.get('path', None)
-                if path is not None:
-                    if path.startswith('s3://'):
-                        endpoint = storage_config.get('endpoint', None)
-                        access_key = storage_config.get('access_key', None)
-                        secret_key = storage_config.get('secret_key', None)
-                        return S3Storage(path, endpoint, access_key,
-                                         secret_key)
-                    elif path.startswith('sftp://'):
-                        return SftpStorage()
-                    else:
-                        return LocalDiskStorage()
-                else:
-                    raise CLIException(ExistCode.NOT_SUPPORTED_STORE_PATH,
-                                       f'unsupported storage path: {path}')
-            else:
-                raise CLIException(
-                    ExistCode.STORAGE_NAME_NOT_EXIST,
-                    f"storage name key `{storage_name}' not found in config")
-        else:
-            raise CLIException(
-                ExistCode.STORAGE_NOT_EXIST,
-                f"storage named `{storage_name}' not found in config")
-
-
-if __name__ == "__main__":
-    s: Storage = StorageBuilder.build_by_name(
-        "s3", {
-            "storage": {
-                "s3": {
-                    "endpoint": "http://10.140.0.94:9800",
-                    "access_key": "ailabminio",
-                    "secret_key": "123123123",
-                    "path": "s3://testdata"
-                }
-            }
-        })
-    b = s.remove_tree("s3://testdata/.dsdl")
-    print(b)
